fix(entity): renaming BaseModel to BaseEntity
