+++ server/src/entities/base/BaseEntity.ts
